amd-ppr: support for RT PPR in 2x1P mode
Add support for Runtime PPR in 2x1P mode.
user should set "SocNum"  to 0 for P0 and 1 for P1
in the RT PPR JSON file.
in 2P mode, BMC always send the request to P0.
in 2x1P mode, BMC use "SocNum" to send the request to
proper SOC.

tested:
- verified in Morocco

Signed-off-by: modolaty <mohsen.dolaty@amd.com>